changed sum_stdin script to work with float instead of integers

+++ sum_stdin.py
@@ -4,7 +4,7 @@ res = 0
 stdin_all = sys.stdin.read()
 # print(stdin_all)
 for elem in stdin_all.split("\n"):
-    numer = int(elem) if len(elem) > 0 else 0
+    numer = float(elem) if len(elem) > 0 else 0
     res += numer
 print(res)
